Drupal check script added
This script is used for cycling through Drupal sites on a system (/home/jenkins/workspace in this case as they're dev systems with CI) and generating lists of modules installed on each site. Results are stored as follows:

"results" subdir (created in the dir the script is run from) -> "servername" subdirectories -> sitename(s).txt

The script has logic to find the correct location to run Drush in for each site and also has checks for different Drush versions if applicable / available. You can also specify specific module names to be checked for - useful for checking against security advisories. This does not update the modules; just lets you know that you could be affected.